backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.3.3 # For BaseSettings
python-multipart==0.0.9 # For OAuth2PasswordRequestForm
python-dotenv==1.0.1
websockets
python-multipart